+++ lib/cmdx/log_formatters/key_value.rb
@@ -4,14 +4,15 @@ module CMDx
   module LogFormatters
     class KeyValue
 
-      def call(_severity, _time, _progname, message)
-        if message.is_a?(Result)
-          message = message.to_h.map do |k, v|
-            "#{k}=#{v}"
-          end.join(" ")
-        end
+      def call(severity, time, task, message)
+        m = LoggerSerializer.call(severity, time, task, message).merge!(
+          severity:,
+          pid: Process.pid,
+          timestamp: Utils::LogTimestamp.call(time.utc)
+        )
 
-        message << "\n"
+        m = m.map { |k, v| "#{k}=#{v}" }.join(" ") if m.is_a?(Hash)
+        m << "\n"
       end
 
     end