+++ spec/cmdx/log_formatters/json_spec.rb
@@ -11,15 +11,18 @@ RSpec.describe CMDx::LogFormatters::Json do
         {"index":0,
         "run_id":"018c2b95-b764-7615-a924-cc5b910ed1e5",
         "type":"Task",
-        "class":"SimulationTask",
+        "task":"SimulationTask",
         "id":"018c2b95-b764-7615-a924-cc5b910ed1e5",
+        "tags":[],
         "state":"complete",
         "status":"success",
         "outcome":"success",
         "metadata":{},
         "runtime":0,
-        "tags":[],
-        "pid":3784}
+        "origin":"CMDx",
+        "severity":"INFO",
+        "pid":3784,
+        "timestamp":"2022-07-17T18:43:15.000000"}
       LINE
     end
   end