+++ spec/cmdx/log_formatters/line_spec.rb
@@ -9,19 +9,19 @@ RSpec.describe CMDx::LogFormatters::Line do
         local_io = LogFormatterHelpers.simulation_output(described_class, :success)
 
         expect(local_io).to match_log(<<~LINE.tr("\n", " "))
-          I, [2022-07-17T18:43:15.000000 #3784] INFO -- CMDx:
+          I, [2022-07-17T18:43:15.000000 #3784] INFO -- SimulationTask:
           index=0
           run_id=018c2b95-b764-7615-a924-cc5b910ed1e5
           type=Task
-          class=SimulationTask
+          task=SimulationTask
           id=018c2b95-b764-7615-a924-cc5b910ed1e5
+          tags=[]
           state=complete
           status=success
           outcome=success
           metadata={}
           runtime=0
-          tags=[]
-          pid=3784
+          origin=CMDx
         LINE
       end
     end
@@ -31,19 +31,19 @@ RSpec.describe CMDx::LogFormatters::Line do
         local_io = LogFormatterHelpers.simulation_output(described_class, :skipped)
 
         expect(local_io).to match_log(<<~LINE.tr("\n", " "))
-          W, [2022-07-17T18:43:15.000000 #3784] WARN -- CMDx:
+          W, [2022-07-17T18:43:15.000000 #3784] WARN -- SimulationTask:
           index=0
           run_id=018c2b95-b764-7615-a924-cc5b910ed1e5
           type=Task
-          class=SimulationTask
+          task=SimulationTask
           id=018c2b95-b764-7615-a924-cc5b910ed1e5
+          tags=[]
           state=interrupted
           status=skipped
           outcome=skipped
           metadata={}
           runtime=0
-          tags=[]
-          pid=3784
+          origin=CMDx
         LINE
       end
     end
@@ -53,19 +53,19 @@ RSpec.describe CMDx::LogFormatters::Line do
         local_io = LogFormatterHelpers.simulation_output(described_class, :failed)
 
         expect(local_io).to match_log(<<~LINE.tr("\n", " "))
-          E, [2022-07-17T18:43:15.000000 #3784] ERROR -- CMDx:
+          E, [2022-07-17T18:43:15.000000 #3784] ERROR -- SimulationTask:
           index=0
           run_id=018c2b95-b764-7615-a924-cc5b910ed1e5
           type=Task
-          class=SimulationTask
+          task=SimulationTask
           id=018c2b95-b764-7615-a924-cc5b910ed1e5
+          tags=[]
           state=interrupted
           status=failed
           outcome=failed
           metadata={}
           runtime=0
-          tags=[]
-          pid=3784
+          origin=CMDx
         LINE
       end
     end
@@ -76,39 +76,39 @@ RSpec.describe CMDx::LogFormatters::Line do
 
         if RubyVersionHelpers.atleast?(3.4)
           expect(local_io).to match_log(<<~LINE.tr("\n", " "))
-            E, [2022-07-17T18:43:15.000000 #3784] ERROR -- CMDx:
+            E, [2022-07-17T18:43:15.000000 #3784] ERROR -- SimulationTask:
             index=0
             run_id=018c2b95-b764-7615-a924-cc5b910ed1e5
             type=Task
-            class=SimulationTask
+            task=SimulationTask
             id=018c2b95-b764-7615-a924-cc5b910ed1e5
+            tags=[]
             state=interrupted
             status=failed
             outcome=interrupted
             metadata={}
             runtime=0
-            tags=[]
-            pid=3784
-            caused_failure={index: 1, run_id: "018c2b95-b764-7615-a924-cc5b910ed1e5", type: "Task", class: "SimulationTask", id: "018c2b95-b764-7615-a924-cc5b910ed1e5", state: "interrupted", status: "failed", outcome: "failed", metadata: {}, runtime: 0, tags: [], pid: 3784}
-            threw_failure={index: 1, run_id: "018c2b95-b764-7615-a924-cc5b910ed1e5", type: "Task", class: "SimulationTask", id: "018c2b95-b764-7615-a924-cc5b910ed1e5", state: "interrupted", status: "failed", outcome: "failed", metadata: {}, runtime: 0, tags: [], pid: 3784}
+            caused_failure={index: 1, run_id: "018c2b95-b764-7615-a924-cc5b910ed1e5", type: "Task", task: "SimulationTask", id: "018c2b95-b764-7615-a924-cc5b910ed1e5", tags: [], state: "interrupted", status: "failed", outcome: "failed", metadata: {}, runtime: 0}
+            threw_failure={index: 1, run_id: "018c2b95-b764-7615-a924-cc5b910ed1e5", type: "Task", task: "SimulationTask", id: "018c2b95-b764-7615-a924-cc5b910ed1e5", tags: [], state: "interrupted", status: "failed", outcome: "failed", metadata: {}, runtime: 0}
+            origin=CMDx
           LINE
         else
           expect(local_io).to match_log(<<~LINE.tr("\n", " "))
-            E, [2022-07-17T18:43:15.000000 #3784] ERROR -- CMDx:
+            E, [2022-07-17T18:43:15.000000 #3784] ERROR -- SimulationTask:
             index=0
             run_id=018c2b95-b764-7615-a924-cc5b910ed1e5
             type=Task
-            class=SimulationTask
+            task=SimulationTask
             id=018c2b95-b764-7615-a924-cc5b910ed1e5
+            tags=[]
             state=interrupted
             status=failed
             outcome=interrupted
             metadata={}
             runtime=0
-            tags=[]
-            pid=3784
-            caused_failure={:index=>1, :run_id=>"018c2b95-b764-7615-a924-cc5b910ed1e5", :type=>"Task", :class=>"SimulationTask", :id=>"018c2b95-b764-7615-a924-cc5b910ed1e5", :state=>"interrupted", :status=>"failed", :outcome=>"failed", :metadata=>{}, :runtime=>0, :tags=>[], :pid=>3784}
-            threw_failure={:index=>1, :run_id=>"018c2b95-b764-7615-a924-cc5b910ed1e5", :type=>"Task", :class=>"SimulationTask", :id=>"018c2b95-b764-7615-a924-cc5b910ed1e5", :state=>"interrupted", :status=>"failed", :outcome=>"failed", :metadata=>{}, :runtime=>0, :tags=>[], :pid=>3784}
+            caused_failure={:index=>1, :run_id=>"018c2b95-b764-7615-a924-cc5b910ed1e5", :type=>"Task", :task=>"SimulationTask", :id=>"018c2b95-b764-7615-a924-cc5b910ed1e5", :tags=>[], :state=>"interrupted", :status=>"failed", :outcome=>"failed", :metadata=>{}, :runtime=>0}
+            threw_failure={:index=>1, :run_id=>"018c2b95-b764-7615-a924-cc5b910ed1e5", :type=>"Task", :task=>"SimulationTask", :id=>"018c2b95-b764-7615-a924-cc5b910ed1e5", :tags=>[], :state=>"interrupted", :status=>"failed", :outcome=>"failed", :metadata=>{}, :runtime=>0}
+            origin=CMDx
           LINE
         end
       end
@@ -120,39 +120,39 @@ RSpec.describe CMDx::LogFormatters::Line do
 
         if RubyVersionHelpers.atleast?(3.4)
           expect(local_io).to match_log(<<~LINE.tr("\n", " "))
-            E, [2022-07-17T18:43:15.000000 #3784] ERROR -- CMDx:
+            E, [2022-07-17T18:43:15.000000 #3784] ERROR -- SimulationTask:
             index=0
             run_id=018c2b95-b764-7615-a924-cc5b910ed1e5
             type=Task
-            class=SimulationTask
+            task=SimulationTask
             id=018c2b95-b764-7615-a924-cc5b910ed1e5
+            tags=[]
             state=interrupted
             status=failed
             outcome=interrupted
             metadata={}
             runtime=0
-            tags=[]
-            pid=3784
-            caused_failure={index: 2, run_id: "018c2b95-b764-7615-a924-cc5b910ed1e5", type: "Task", class: "SimulationTask", id: "018c2b95-b764-7615-a924-cc5b910ed1e5", state: "interrupted", status: "failed", outcome: "failed", metadata: {}, runtime: 0, tags: [], pid: 3784}
-            threw_failure={index: 1, run_id: "018c2b95-b764-7615-a924-cc5b910ed1e5", type: "Task", class: "SimulationTask", id: "018c2b95-b764-7615-a924-cc5b910ed1e5", state: "interrupted", status: "failed", outcome: "interrupted", metadata: {}, runtime: 0, tags: [], pid: 3784}
+            caused_failure={index: 2, run_id: "018c2b95-b764-7615-a924-cc5b910ed1e5", type: "Task", task: "SimulationTask", id: "018c2b95-b764-7615-a924-cc5b910ed1e5", tags: [], state: "interrupted", status: "failed", outcome: "failed", metadata: {}, runtime: 0}
+            threw_failure={index: 1, run_id: "018c2b95-b764-7615-a924-cc5b910ed1e5", type: "Task", task: "SimulationTask", id: "018c2b95-b764-7615-a924-cc5b910ed1e5", tags: [], state: "interrupted", status: "failed", outcome: "interrupted", metadata: {}, runtime: 0}
+            origin=CMDx
           LINE
         else
           expect(local_io).to match_log(<<~LINE.tr("\n", " "))
-            E, [2022-07-17T18:43:15.000000 #3784] ERROR -- CMDx:
+            E, [2022-07-17T18:43:15.000000 #3784] ERROR -- SimulationTask:
             index=0
             run_id=018c2b95-b764-7615-a924-cc5b910ed1e5
             type=Task
-            class=SimulationTask
+            task=SimulationTask
             id=018c2b95-b764-7615-a924-cc5b910ed1e5
+            tags=[]
             state=interrupted
             status=failed
             outcome=interrupted
             metadata={}
             runtime=0
-            tags=[]
-            pid=3784
-            caused_failure={:index=>2, :run_id=>"018c2b95-b764-7615-a924-cc5b910ed1e5", :type=>"Task", :class=>"SimulationTask", :id=>"018c2b95-b764-7615-a924-cc5b910ed1e5", :state=>"interrupted", :status=>"failed", :outcome=>"failed", :metadata=>{}, :runtime=>0, :tags=>[], :pid=>3784}
-            threw_failure={:index=>1, :run_id=>"018c2b95-b764-7615-a924-cc5b910ed1e5", :type=>"Task", :class=>"SimulationTask", :id=>"018c2b95-b764-7615-a924-cc5b910ed1e5", :state=>"interrupted", :status=>"failed", :outcome=>"interrupted", :metadata=>{}, :runtime=>0, :tags=>[], :pid=>3784}
+            caused_failure={:index=>2, :run_id=>"018c2b95-b764-7615-a924-cc5b910ed1e5", :type=>"Task", :task=>"SimulationTask", :id=>"018c2b95-b764-7615-a924-cc5b910ed1e5", :tags=>[], :state=>"interrupted", :status=>"failed", :outcome=>"failed", :metadata=>{}, :runtime=>0}
+            threw_failure={:index=>1, :run_id=>"018c2b95-b764-7615-a924-cc5b910ed1e5", :type=>"Task", :task=>"SimulationTask", :id=>"018c2b95-b764-7615-a924-cc5b910ed1e5", :tags=>[], :state=>"interrupted", :status=>"failed", :outcome=>"interrupted", :metadata=>{}, :runtime=>0}
+            origin=CMDx
           LINE
         end
       end