+++ spec/cmdx/result_serializer_spec.rb
@@ -14,15 +14,14 @@ RSpec.describe CMDx::ResultSerializer do
           index: 0,
           run_id: "018c2b95-b764-7615-a924-cc5b910ed1e5",
           type: "Task",
-          class: "SimulationTask",
+          task: "SimulationTask",
           id: "018c2b95-b764-7615-a924-cc5b910ed1e5",
           outcome: CMDx::Result::SUCCESS,
           state: CMDx::Result::COMPLETE,
           status: CMDx::Result::SUCCESS,
           metadata: {},
           runtime: 0,
-          tags: [],
-          pid: 3784
+          tags: []
         )
       end
     end
@@ -35,7 +34,7 @@ RSpec.describe CMDx::ResultSerializer do
           index: 0,
           run_id: "018c2b95-b764-7615-a924-cc5b910ed1e5",
           type: "Task",
-          class: "SimulationTask",
+          task: "SimulationTask",
           id: "018c2b95-b764-7615-a924-cc5b910ed1e5",
           outcome: CMDx::Result::INTERRUPTED,
           state: CMDx::Result::INTERRUPTED,
@@ -43,14 +42,12 @@ RSpec.describe CMDx::ResultSerializer do
           metadata: {},
           runtime: 0,
           tags: [],
-          pid: 3784,
           caused_failure: {
-            class: "SimulationTask",
+            task: "SimulationTask",
             id: "018c2b95-b764-7615-a924-cc5b910ed1e5",
             index: 2,
             metadata: {},
             outcome: CMDx::Result::FAILED,
-            pid: 3784,
             runtime: 0,
             run_id: "018c2b95-b764-7615-a924-cc5b910ed1e5",
             state: CMDx::Result::INTERRUPTED,
@@ -59,12 +56,11 @@ RSpec.describe CMDx::ResultSerializer do
             type: "Task"
           },
           threw_failure: {
-            class: "SimulationTask",
+            task: "SimulationTask",
             id: "018c2b95-b764-7615-a924-cc5b910ed1e5",
             index: 1,
             metadata: {},
             outcome: CMDx::Result::INTERRUPTED,
-            pid: 3784,
             runtime: 0,
             run_id: "018c2b95-b764-7615-a924-cc5b910ed1e5",
             state: CMDx::Result::INTERRUPTED,