+++ spec/cmdx/run_inspector_spec.rb
@@ -10,7 +10,7 @@ RSpec.describe CMDx::RunInspector do
       expect(run.to_s).to eq(<<~TEXT.chomp)
         Run: 018c2b95-b764-7615-a924-cc5b910ed1e5
         =======================================================
-        SimulationTask: type=Task index=0 id=018c2b95-b764-7615-a924-cc5b910ed1e5 state=complete status=success outcome=success metadata={} tags=[] pid=3784 runtime=0
+        SimulationTask: type=Task index=0 id=018c2b95-b764-7615-a924-cc5b910ed1e5 state=complete status=success outcome=success metadata={} tags=[] runtime=0
         =======================================================
         state=complete status=success outcome=success runtime=0
       TEXT